+++ components/core/ui/decompose/src/main/java/com/flipperdevices/ui/decompose/DecomposeComponent.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.ui.decompose
+
+import androidx.compose.runtime.Composable
+
+interface DecomposeComponent {
+    @Composable
+    fun Render()
+}