+++ components/filemanager/api/src/main/java/com/flipperdevices/filemanager/api/navigation/FileManagerEntry.kt
@@ -1,13 +1,11 @@
 package com.flipperdevices.filemanager.api.navigation
 
-import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
+import com.flipperdevices.core.ui.navigation.ComposableFeatureEntry
 import com.flipperdevices.core.ui.navigation.FeatureScreenRootRoute
 
-private const val DEFAULT_PATH = "/"
-
-interface FileManagerEntry : AggregateFeatureEntry {
+interface FileManagerEntry : ComposableFeatureEntry {
     override val ROUTE: FeatureScreenRootRoute
         get() = FeatureScreenRootRoute.FILE_MANAGER
 
-    fun fileManagerDestination(path: String = DEFAULT_PATH): String
+    fun fileManagerDestination(): String
 }